% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-quasar2CR.R
\name{coef.quasar2CR}
\alias{coef.quasar2CR}
\title{Extract regression coefficients from a quasar2CR fit.}
\usage{
\method{coef}{quasar2CR}(object, id = NULL, ...)
}
\arguments{
\item{object}{a \code{quasar2CR} object.}

\item{id}{optional character; if supplied, return only this identifier's
coefficients.  Otherwise return the named list of all.}

\item{...}{ignored.}
}
\value{
numeric vector (one SNP) or named list (all SNPs).
}
\description{
Extract regression coefficients from a quasar2CR fit.
}
