% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.rho.R
\name{comp.rho}
\alias{comp.rho}
\title{rho calculation}
\usage{
comp.rho(ref, alt, eps)
}
\arguments{
\item{ref}{integer. reference allele count}

\item{alt}{integer. alternate allele count}

\item{eps}{float. error estimate in [0,1]}
}
\value{
rho float. elstimate of alternate allele frequency
}
\description{
given reference and alternate read counts, & an error estimate, return the 
estimated alternate alle frequency
}
