% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitQuasar2.R
\name{fitQuasar2}
\alias{fitQuasar2}
\title{fitQuasar2 Fits Beta-Binomial Logistic Regression Model on multiple groups of SNP and conditions}
\usage{
fitQuasar2(
  dd,
  model,
  nbreaks = 20,
  M = 20,
  eps = 0.001,
  tolM = 0.01,
  tolCoeff = 0.02,
  tolEps = 2e-04,
  maxIt = 100
)
}
\arguments{
\item{dd}{Data frame containing the data to be modeled. The data frame should include the
predictor variables specified in the model formula, along with the following columns:
- `identifier` : the betabinomial model will be fit based on grouping variable. 
For MPRA experiments this would be a construct ID, for ASE/cASE on individual SNPs it 
will be the SNP id combined with an individual ID. It is important NOT to mix individuals
unless some phasing is done beforehand flipping R/A accordingly. 
- `R`: Number of successes (e.g., reads matching the reference allele).
- `A`: Number of failures (e.g., reads matching the alternate allele).
- `offset`: Offset value if applicable, default is 0. Useful for MPRA (`qlogis(DNA_prop)`)
   This is optional, and it does not need to be set.}

\item{model}{Formula specifying the model to be fitted, like ~ Batch + Treatment. Variables have to
present in the `dd` object as additional columns and properly formated as factors or continuous 
variables.}

\item{nbreaks}{Number of breaks to divide the overall expression mean of the `identifier` with `R+A`
which is the baseMean. For each break an overdispersion parameter M will be calculated.}

\item{M}{Initial values for the overdispersion parameter (default 20)}

\item{eps}{Initial value to account for sequencing errors. Small constant added to probabilities 
to prevent them from being 0 or 1, default is 0.001.}
}
\value{
A list containing two elements: `res`, a tibble with the fitted coefficients, standard errors,
        statistical significance, and other diagnostic measures; and `dd`, the modified data frame
        with additional columns for fitted probabilities.
}
\description{
This function fits a beta-binomial logistic regression model to the data based on a specified
model formula. It is ideal for cases with allele-specific expression (ASE) data that exhibits
overdispersion relative to a standard binomial model. The function employs gradient descent 
optimization techniques to estimate the best fitting model parameters of the logistic model. 
This function is designed to be used on data that has been grouped by SNP:individual sharing
the overdispersion parameter M, which is fit on multiple quantile breaks on number of reads.
}
\examples{
\dontrun{
  # Assuming `mydat` is a data frame with required columns and model predictors
  model_formula <- ~ batch + Tr
  initial_b <- rep(0, length = ncol(model.matrix(model_formula, mydat)))
  fitted_model <- fitBetaBinomialLogistic(mydat, model_formula, b = initial_b)
}
}
\seealso{
\code{\link{fitBetaBinomialLogistic}}: Log-likelihood function for beta-binomial logistic regression.
}
