% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitQuasarMpra.R
\name{fitQuasarMpra}
\alias{fitQuasarMpra}
\title{Title fitQuasarMpra}
\usage{
fitQuasarMpra(ref, alt, prop = 0.5, eps = 0.001, nbreaks = 10, M = 0)
}
\arguments{
\item{ref}{reference read count}

\item{alt}{alternate read count}

\item{prop}{reference DNA proportion (def.=0.5)}

\item{eps}{base-calling error rate (def.=0.001)}

\item{nbreaks}{number of breaks of coverage to estimate dispersion (def.=10)}
}
\value{
data.frame with the following fields: 
- bin total coverage bin used
- betas.beta.binom  logit transfomation of the allelic skew
- betas_se standard error of beta 
- betas_z zscore for beta-plogis(propr) 
- pval3 p.value 
- padj_quasar BH adjusted p.value
}
\description{
For an MPRA experiment assess allele specific expression.
}
