% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLogLikBetaBinomialEps.R
\name{gLogLikBetaBinomialEps}
\alias{gLogLikBetaBinomialEps}
\title{Gradient of Log-Likelihood for Beta-Binomial Model to Optimize Epsilon Using Estimated ASE Proportion}
\usage{
gLogLikBetaBinomialEps(eps, M, p, R, A)
}
\arguments{
\item{eps}{Probability of base call error to be estimated}

\item{M}{Concentration parameter of the beta distribution ($\alpha + \beta$).}

\item{p}{Estimated proportion of the reference allele expression}

\item{R}{Number of reads matching the reference allele.}

\item{A}{Number of reads matching the alternate allele.}
}
\value{
Gradient of the negative log-likelihood with respect to epsilon.
}
\description{
This function calculates the gradient of the log-likelihood for a beta-binomial model, with a
focus on optimizing the epsilon parameter. It is useful in optimization algorithms that require 
gradient information.
}
\examples{
M <- 10
p <- 0.5  # Estimated proportion
R <- 3
A <- 7
eps <- 0.001
gLogLikBetaBinomialEps(eps, M, p, R, A)
}
