% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLogLikBetaBinomialLogistic.R
\name{gLogLikBetaBinomialLogistic}
\alias{gLogLikBetaBinomialLogistic}
\title{Gradient of Log-Likelihood for Beta-Binomial Logistic Regression to Assess ASE}
\usage{
gLogLikBetaBinomialLogistic(
  b,
  X,
  M,
  R,
  A,
  offset = 0,
  eps = 0.001,
  lambda = 0.001
)
}
\arguments{
\item{b}{Numeric vector of coefficients.}

\item{X}{Design matrix of predictors.}

\item{M}{Concentration parameter of the beta distribution ($\alpha + \beta$).}

\item{R}{Number of reads matching the reference allele.}

\item{A}{Number of reads matching the alternate allele.}

\item{offset}{Adjustment for the offset in logistic regression, default is 0.0,
useful when MPRA expected ASE is not 0.5.}

\item{eps}{Probability of base call error, default is 0.001.}

\item{lambda}{Regularization parameter for L2 regularization, default is 0.001.}
}
\value{
A numeric vector representing the gradient of the negative log-likelihood.
}
\description{
This function computes the gradient of the log-likelihood for a beta-binomial
logistic regression model, useful for detecting allele-specific expression (ASE) changes
across different conditions or contexts. This function is often used for optimization purposes
in maximum likelihood estimation, helping to find the parameters b values that maximize
the log-likelihood function. It also adjusts for potential offsets and includes an L2 
regularization term.
}
\examples{
b <- c(0.5, -0.3)
X <- matrix(c(1, 2, 3, 1, 5, 2), ncol = 2)
M <- 10
R <- c(3, 6, 2)
A <- c(5, 5, 5)
gLogLikBetaBinomialLogistic(b, X, M, R, A)
}
