% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gLogLikBetaBinomialM.R
\name{gLogLikBetaBinomialM}
\alias{gLogLikBetaBinomialM}
\title{Gradient of Log-Likelihood for Beta-Binomial Model M Estimation Using Estimated ASE Proportion}
\usage{
gLogLikBetaBinomialM(M, p, R, A, eps = 0.001)
}
\arguments{
\item{M}{Concentration parameter of the beta distribution ($\alpha + \beta$).}

\item{p}{Estimated proportion of the reference allele expression, as derived from best fit.}

\item{R}{Number of reads matching the reference allele.}

\item{A}{Number of reads matching the alternate allele.}

\item{eps}{Probability of base call error, default is 0.001.}
}
\value{
Numeric vector representing the gradient of the negative log-likelihood.
}
\description{
This function calculates the gradient of the log-likelihood for a beta-binomial model,
which is useful for optimization algorithms aiming to find the best parameter estimates for M.
It is designed to work with allele-specific expression (ASE) data where the proportion
of allele expression has been estimated. The function helps in adjusting the concentration
parameter M based on the gradient descent method or similar optimization approaches.
}
\examples{
M <- 10
p <- 0.5  # Best fit proportion estimate
R <- 3
A <- 7
gLogLikBetaBinomialM(M, p, R, A)
}
