% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBetaBinomial2.R
\name{logLikBetaBinomial2}
\alias{logLikBetaBinomial2}
\title{log-likelihood of beta-binomial model for heterozygotes}
\usage{
logLikBetaBinomial2(logit.p, D, R, A)
}
\arguments{
\item{logit.p}{logit(rho) , rho value in [0, 1]}

\item{D}{dispersion estimate}

\item{R}{reference read count}

\item{A}{alternate read count}
}
\value{
log likelihood
}
\description{
returns the log likelihood of reference & alternate read count data
given rho, & dispersion.
}
