% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBetaBinomialEps.R
\name{logLikBetaBinomialEps}
\alias{logLikBetaBinomialEps}
\title{Log-Likelihood for Beta-Binomial Model to Optimize Epsilon Using Estimated ASE Proportion}
\usage{
logLikBetaBinomialEps(eps, M, p, R, A)
}
\arguments{
\item{eps}{Probability of base call error to be estimated}

\item{M}{Concentration parameter of the beta distribution ($\alpha + \beta$).}

\item{p}{Estimated proportion of the reference allele expression}

\item{R}{Number of reads matching the reference allele.}

\item{A}{Number of reads matching the alternate allele.}
}
\value{
Numeric value representing the negative log-likelihood.
}
\description{
This function calculates the log-likelihood for a beta-binomial model, focusing on optimizing
the epsilon parameter. Epsilon is probability of base call error when
analyzing allele-specific expression (ASE).
}
\examples{
M <- 10
p <- 0.5  # Estimated proportion
R <- 3
A <- 7
eps <- 0.001
logLikBetaBinomialEps(eps, M, p, R, A)
}
