% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBetaBinomialLogistic.R
\name{logLikBetaBinomialLogistic}
\alias{logLikBetaBinomialLogistic}
\title{Log-likelihood for Beta-Binomial Logistic Regression to Assess ASE}
\usage{
logLikBetaBinomialLogistic(
  b,
  X,
  M,
  R,
  A,
  offset = 0,
  eps = 0.001,
  lambda = 0.001
)
}
\arguments{
\item{b}{Numeric vector of coefficients.}

\item{X}{Design matrix of predictors.}

\item{M}{Concentration parameter of the beta distribution ($\alpha + \beta$).}

\item{R}{Number of reads matching the reference allele.}

\item{A}{Number of reads matching the alternate allele.}

\item{offset}{Adjustment for the offset in logistic regression, default is 0.0,
useful when MPRA expected ASE is not 0.5.}

\item{eps}{Probability of base call error, default is 0.001.}

\item{lambda}{Regularization parameter for L2 regularization, default is 0.001.}
}
\value{
Numeric value representing the negative log-likelihood.
}
\description{
This function calculates the log-likelihood for a beta-binomial logistic regression model,
which is useful for detecting changes in allele-specific expression (ASE) across different
conditions or contexts specified by a design matrix. It adjusts for potential offsets if ASE
is not 0.5 under the null hypothesis, which is useful for MPRA applications. Additionally,
an L2 regularization term can be applied to control overfitting.
}
\examples{
b <- c(0.5, -0.3)
X <- matrix(c(1, 2, 3, 1, 5, 2), ncol = 2)
M <- 10
R <- c(3, 6, 2)
A <- c(5, 5, 5)
logLikBetaBinomialLogistic(b, X, M, R, A)
}
