% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBetaBinomialM.R
\name{logLikBetaBinomialM}
\alias{logLikBetaBinomialM}
\title{Log-Likelihood for Beta-Binomial Model To Estimate M Using Estimated ASE Proportion Parameter}
\usage{
logLikBetaBinomialM(M, p, R, A, eps = 0.001)
}
\arguments{
\item{M}{Concentration parameter of the beta distribution ($\alpha + \beta$)}

\item{p}{Estimated proportion of the reference allele expression, as derived from best fit.}

\item{R}{Number of reads matching the reference allele.}

\item{A}{Number of reads matching the alternate allele.}

\item{eps}{Probability of base call error, default is 0.001.}
}
\value{
Numeric value representing the negative log-likelihood.
}
\description{
This function calculates the log-likelihood for a beta-binomial model, suitable for analyzing
allele-specific expression (ASE) where the proportion of allele expression is estimated from
data. It is designed to estimate the concentration parameter and using best fit and the 
the likelihood of observing the given number of reference and alternate allele reads.
}
\examples{
M <- 10
p <- 0.5  # Best fit proportion estimate
R <- 3
A <- 7
logLikBetaBinomialM(M, p, R, A)
}
