% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepForGenotyping.R
\name{PrepForGenotyping}
\alias{PrepForGenotyping}
\title{PrepForGenotyping}
\usage{
PrepForGenotyping(ase.dat, min.coverage, dampen.priors = TRUE)
}
\arguments{
\item{ase.dat}{data from UnionExtractFields.R}

\item{min.coverage}{minimumm coverage across all combined sample for locus inclusion}

\item{dampen.priors}{prevents underflow}
}
\value{
a list with reference allele counts, alternate allele counts, a matrix of gt priors, and annotations
}
\description{
PrepForGenotyping
}
