% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-quasar2CR.R
\name{vcov.quasar2CR}
\alias{vcov.quasar2CR}
\title{Extract Wald variance-covariance matrices from a quasar2CR fit.}
\usage{
\method{vcov}{quasar2CR}(object, id = NULL, ...)
}
\arguments{
\item{object}{a \code{quasar2CR} object.}

\item{id}{optional character; identifier of a single SNP.}

\item{...}{ignored.}
}
\value{
matrix (one SNP) or named list (all SNPs).
}
\description{
Extract Wald variance-covariance matrices from a quasar2CR fit.
}
